Move bench command to the separate file
